CsvOptions enhancements
